restructure routes -> router
